requirements.txt
python-lzo
fusepy